refactor(certificates): move certificate to utils
